requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pypdf==4.3.1
python-dotenv==1.0.1
nltk==3.9.1
requests==2.32.5
openai==1.50.2
pydantic==2.9.2
